refactor: split runner daemon refresh helpers
